feat(operator): wire crane@smd.services as inbound email address with allow-list gating
Adds the email-reply skill so the SMD Operator can receive and reply to
email sent directly to crane@smd.services. Polls on a 15-min schedule
(7am–8pm), gates on scope.inbound_allow_from, falls back to draft when
workspace_gmail_send is not yet available (Wave 2 prereq).

New sections-scope.ts extracted from sections-other.ts to satisfy the
500-line max-lines ceiling after inbound_allow_from was added to Scope.
