+++ src/lib/operator/customer-yaml/sections-other.ts
@@ -1,7 +1,7 @@
 /**
- * Remaining section validators: scope, escalation, memory, and the
+ * Remaining section validators: escalation, memory, and the
  * optional voice_library / business_hours / logging / pause / observability
- * blocks.
+ * blocks. Scope validation lives in sections-scope.ts.
  */
 
 import {
@@ -17,61 +17,11 @@ import {
   type Memory,
   type MemoryRetention,
   type Pause,
-  type Scope,
   type ValidationError,
   type Vertical,
   type VoiceLibrary,
 } from './types'
-import {
-  isPlainObject,
-  optionalNonEmptyString,
-  optionalStringList,
-  requireStringList,
-} from './helpers'
-
-export function checkScope(root: Record<string, unknown>, errors: ValidationError[]): Scope {
-  const raw = root['scope']
-  if (raw === undefined || raw === null) {
-    errors.push({ code: 'MissingField', path: 'scope', message: 'scope is required' })
-    return emptyScope()
-  }
-  if (!isPlainObject(raw)) {
-    errors.push({ code: 'TypeMismatch', path: 'scope', message: 'scope must be an object' })
-    return emptyScope()
-  }
-  return {
-    email_folders_visible: requireStringList(
-      raw,
-      'email_folders_visible',
-      'scope.email_folders_visible',
-      errors
-    ),
-    email_folders_blind: requireStringList(
-      raw,
-      'email_folders_blind',
-      'scope.email_folders_blind',
-      errors
-    ),
-    email_keyword_blocks: requireStringList(
-      raw,
-      'email_keyword_blocks',
-      'scope.email_keyword_blocks',
-      errors
-    ),
-    domain_blocks: requireStringList(raw, 'domain_blocks', 'scope.domain_blocks', errors),
-    matter_blocks: optionalStringList(raw, 'matter_blocks', 'scope.matter_blocks', errors),
-  }
-}
-
-function emptyScope(): Scope {
-  return {
-    email_folders_visible: [],
-    email_folders_blind: [],
-    email_keyword_blocks: [],
-    domain_blocks: [],
-    matter_blocks: [],
-  }
-}
+import { isPlainObject, optionalNonEmptyString, requireStringList } from './helpers'
 
 export function checkEscalation(
   root: Record<string, unknown>,
